Multi Database Connection/pyproject.toml
[project]
name = "multi-database-connection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.6.0",
    "psycopg2-binary>=2.9.11",
]
